third_party/WebKit/LayoutTests/crypto/subtle/clone-hkdfKey.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests structured cloning of HKDF keys");

jsTestIsAsync = true;

// Tests the 18 permutations of keys generated by:
//   kPossibleHashAlgorithms x kPossibleKeyUsages x kPossibleKeyData
//
// For practical reasons these tests are not exhaustive.

var k128BitData = "30112233445566778899aabbccddeeff"
var k256BitData = "00112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f";

var kPossibleHashAlgorithms = ['SHA-1', 'SHA-256', 'SHA-512'];
var kPossibleKeyUsages = [['deriveBits'], ['deriveKey'], ['deriveKey', 'deriveBits']];
var kPossibleKeyData = [
    k128BitData,
    k256BitData
];

function runTest(hashName, keyUsages, keyData)
{
    var importData = hexStringToUint8Array(keyData);
    var importAlgorithm = { name: 'HKDF', hash: {name: hashName } };

    var results = {};

    var extractable = false;
    return crypto.subtle.importKey('raw', importData, importAlgorithm, extractable, keyUsages).then(function(importedKey) {
        results.importedKey = importedKey;
        importedKey.extraProperty = 'hi';
        return cloneKey(importedKey);
    }).then(function(result) {
        results.clonedKey = result;

        importedKey = results.importedKey;
        clonedKey = results.clonedKey;

        shouldEvaluateAs("importedKey.extraProperty", "hi");
        shouldEvaluateAs("importedKey.type", "secret");
        shouldEvaluateAs("importedKey.extractable", extractable);
        shouldEvaluateAs("importedKey.algorithm.name", "HKDF");
        shouldEvaluateAs("importedKey.usages.join(',')", keyUsages.join(","));

        shouldNotBe("importedKey", "clonedKey");

        shouldBeUndefined("clonedKey.extraProperty");
        shouldEvaluateAs("clonedKey.type", "secret");
        shouldEvaluateAs("clonedKey.extractable", extractable);
        shouldEvaluateAs("clonedKey.algorithm.name", "HKDF");
        shouldEvaluateAs("clonedKey.usages.join(',')", keyUsages.join(","));

        logSerializedKey(importedKey);

        debug("");
    });
}

var lastPromise = Promise.resolve(null);

kPossibleHashAlgorithms.forEach(function(hashName) {
    kPossibleKeyUsages.forEach(function(keyUsages) {
        kPossibleKeyData.forEach(function(keyData) {
            lastPromise = lastPromise.then(runTest.bind(null, hashName, keyUsages, keyData));
        });
    });
});

lastPromise.then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
