third_party/WebKit/LayoutTests/crypto/subtle/clone-hkdfKey-expected.txt

Tests structured cloning of HKDF keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveBits"
Serialized key bytes: 3f004b061080021030112233445566778899aabbccddeeff

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveBits"
Serialized key bytes: 3f004b061080022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey"
Serialized key bytes: 3f004b0610201030112233445566778899aabbccddeeff00

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey"
Serialized key bytes: 3f004b0610202000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f00

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey,deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey,deriveBits"
Serialized key bytes: 3f004b0610a0021030112233445566778899aabbccddeeff

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey,deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey,deriveBits"
Serialized key bytes: 3f004b0610a0022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveBits"
Serialized key bytes: 3f004b061080021030112233445566778899aabbccddeeff

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveBits"
Serialized key bytes: 3f004b061080022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey"
Serialized key bytes: 3f004b0610201030112233445566778899aabbccddeeff00

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey"
Serialized key bytes: 3f004b0610202000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f00

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey,deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey,deriveBits"
Serialized key bytes: 3f004b0610a0021030112233445566778899aabbccddeeff

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey,deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey,deriveBits"
Serialized key bytes: 3f004b0610a0022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveBits"
Serialized key bytes: 3f004b061080021030112233445566778899aabbccddeeff

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveBits"
Serialized key bytes: 3f004b061080022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey"
Serialized key bytes: 3f004b0610201030112233445566778899aabbccddeeff00

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey"
Serialized key bytes: 3f004b0610202000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f00

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey,deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey,deriveBits"
Serialized key bytes: 3f004b0610a0021030112233445566778899aabbccddeeff

PASS importedKey.extraProperty is "hi"
PASS importedKey.type is "secret"
PASS typeof importedKey.extractable is 'boolean'
PASS importedKey.extractable is false
PASS importedKey.algorithm.name is "HKDF"
PASS importedKey.usages.join(',') is "deriveKey,deriveBits"
PASS importedKey is not clonedKey
PASS clonedKey.extraProperty is undefined.
PASS clonedKey.type is "secret"
PASS typeof clonedKey.extractable is 'boolean'
PASS clonedKey.extractable is false
PASS clonedKey.algorithm.name is "HKDF"
PASS clonedKey.usages.join(',') is "deriveKey,deriveBits"
Serialized key bytes: 3f004b0610a0022000112233445546778899aabbccddeeff000102030405060708090a0b0c0d0e0f

PASS successfullyParsed is true

TEST COMPLETE
